\name{needleman.wunsch-package}
\alias{needleman.wunsch-package}
\alias{needleman.wunsch}
\docType{package}
\title{
Needleman-Wunsch score
}
\description{
Computes the Needleman-Wunsch between to nucleotide sequences.
}
\details{
\tabular{ll}{
Package: \tab needleman.wunsch\cr
Type: \tab Package\cr
Version: \tab 1.4\cr
Date: \tab 2012-12-21\cr
License: \tab GPL2\cr
}
}
\author{
Kristian Hovde Liland

Maintainer: Kristian Hovde Liland <kristian.liland@umb.no>
}
\keyword{ package }
\examples{
example(nw)
example(nws)
example(nw.sat.quer)
example(nws.sat.quer)
example(nw.sg)
example(nws.sg)
example(nw.sg.mult)
example(nws.sg.mult)
}
